refactor: changed timeline routes to reflect simpler module structure
