+++ web-client/src/components/NotificationsDrawer/Notification.tsx
@@ -4,7 +4,7 @@ import { useHistory } from 'react-router-dom';
 import { OfferStatus } from 'src/models/offers';
 import { Request } from 'src/models/requests';
 import { User } from 'src/models/users';
-import { TimelineViewLocation } from 'src/modules/timeline/pages/routes/TimelineViewRoute/constants';
+import { TimelineViewLocation } from 'src/modules/timeline/constants';
 import { COLORS } from 'src/theme/colors';
 import styled from 'styled-components';
