--- web-client/src/modules/timeline/pages/routes/TimelineAcceptedViewRoute/constants.tsx
@@ -1,9 +1,0 @@
-import Location from 'react-app-location';
-import * as Yup from 'yup';
-
-export const TimelineAcceptedViewLocation = new Location(
-  '/timeline/accepted/:requestId',
-  {
-    requestId: Yup.string().required(),
-  },
-);
