--- web-client/src/modules/timeline/pages/routes/TimelineViewRoute/TimelineViewRoute.tsx
@@ -1,19 +1,0 @@
-import React, { ReactElement } from 'react';
-import styled from 'styled-components';
-
-import TimelineViewContainer from '../../../containers/TimelineViewContainer/TimelineViewContainer';
-
-const TimelineViewRoute: React.FC<any> = ({ match }): ReactElement => (
-  <Wrapper>
-    <TimelineViewContainer requestId={match.params.requestId} />
-  </Wrapper>
-);
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-`;
-
-export default TimelineViewRoute;
